Revise Commands
- migrate to `click`
- update dependency
- update README
