+++ pdm.lock
@@ -87,6 +87,15 @@ dependencies = [
 ]
 
 [[package]]
+name = "click"
+version = "8.0.4"
+requires_python = ">=3.6"
+summary = "Composable command line interface toolkit"
+dependencies = [
+    "colorama; platform_system == \"Windows\"",
+]
+
+[[package]]
 name = "colorama"
 version = "0.4.4"
 requires_python = ">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*"
@@ -755,7 +764,7 @@ summary = "A formatter for Python code."
 
 [metadata]
 lock_version = "3.1"
-content_hash = "sha256:249c75ae088ee981e5e4733743dfcd5c6224f2281355963305756cbdfea407b1"
+content_hash = "sha256:cefb73b0e10f90b52f69a04cf63916bad7b399ccd26230156a263363874a4efd"
 
 [metadata.files]
 "appnope 0.1.2" = [
@@ -868,6 +877,10 @@ content_hash = "sha256:249c75ae088ee981e5e4733743dfcd5c6224f2281355963305756cbdf
     {file = "cffi-1.15.0-cp39-cp39-win32.whl", hash = "sha256:2a23af14f408d53d5e6cd4e3d9a24ff9e05906ad574822a10563efcef137979a"},
     {file = "cffi-1.15.0-cp39-cp39-win_amd64.whl", hash = "sha256:3773c4d81e6e818df2efbc7dd77325ca0dcb688116050fb2b3011218eda36139"},
     {file = "cffi-1.15.0.tar.gz", hash = "sha256:920f0d66a896c2d99f0adbb391f990a84091179542c205fa53ce5787aff87954"},
+]
+"click 8.0.4" = [
+    {file = "click-8.0.4-py3-none-any.whl", hash = "sha256:6a7a62563bbfabfda3a38f3023a1db4a35978c0abd76f6c9605ecd6554d6d9b1"},
+    {file = "click-8.0.4.tar.gz", hash = "sha256:8458d7b1287c5fb128c90e23381cf99dcde74beaf6c7ff6384ce84d6fe090adb"},
 ]
 "colorama 0.4.4" = [
     {file = "colorama-0.4.4-py2.py3-none-any.whl", hash = "sha256:9f47eda37229f68eee03b24b9748937c7dc3868f906e8ba69fbcbdd3bc5dc3e2"},
