+++ src/app/proxy_server.py
@@ -6,7 +6,6 @@ import logging
 from typing import NoReturn
 from typing import Optional
 
-from . import cfg
 from .base_protocol import BaseTcpProtocol
 from .base_protocol import CypherProtocol
 
@@ -80,32 +79,3 @@ class ProxyServerProtocol(CypherProtocol):
         await remote.close()
         await self.close()
         return
-
-
-def run():
-    """run it"""
-
-    def handle_client(reader, writer):
-        local = ProxyServerProtocol(reader, writer)
-        LOGGER.info(f'new client from: {local.peer}')
-        return asyncio.ensure_future(local.exchange_data())
-
-    async def service(h: str = cfg.proxy_server['host'],
-                      p: int = cfg.proxy_server['port']):
-        return await asyncio.start_server(handle_client,
-                                          host=h,
-                                          port=p,
-                                          ssl=None)
-
-    loop = asyncio.get_event_loop()
-    server = loop.run_until_complete(service())
-
-    for s in server.sockets:
-        LOGGER.info(f'Proxy broker listening on {s.getsockname()}')
-
-    try:
-        loop.run_forever()
-    except KeyboardInterrupt:
-        pass
-    finally:
-        server.close()
